Expand hitbox size to 16-bit width/height

+++ src/mgc/components/hitbox.h
@@ -17,12 +17,12 @@ typedef struct mgc_hitbox {
     mgc_id_t id;
     int8_t x0_ofs;
     int8_t y0_ofs;
-    uint8_t width;
-    uint8_t height;
+    uint16_t width;
+    uint16_t height;
     bool enabled;
 } mgc_hitbox_t;
 
-void hitbox_init(mgc_hitbox_t *hitbox, mgc_id_t id, int8_t x0_ofs, int8_t y0_ofs, uint8_t width, uint8_t height);
+void hitbox_init(mgc_hitbox_t *hitbox, mgc_id_t id, int8_t x0_ofs, int8_t y0_ofs, uint16_t width, uint16_t height);
 void hitbox_set_enabled(mgc_hitbox_t *hitbox, bool enabled);
 bool hitbox_is_enabled(const mgc_hitbox_t *hitbox);
 const mgc_hitbox_t *hitbox_get_by_id(const mgc_hitbox_t *hitbox_array, size_t hitbox_count, mgc_id_t id);
